rework code around request processor
