--- src/classes/JsonProcessor.php
@@ -1,24 +1,0 @@
-<?php
-/**
- * @author Nicolas CARPi <nico-git@deltablot.email>
- * @copyright 2012 Nicolas CARPi
- * @see https://www.elabftw.net Official website
- * @license AGPL-3.0
- * @package elabftw
- */
-declare(strict_types=1);
-
-namespace Elabftw\Elabftw;
-
-use Symfony\Component\HttpFoundation\Request;
-
-/**
- * Process a JSON payload sent by POST
- */
-class JsonProcessor extends AbstractProcessor
-{
-    protected function process(Request $request): void
-    {
-        $this->processJson((string) $request->getContent());
-    }
-}
