+++ src/interfaces/ProcessorInterface.php
@@ -1,4 +1,4 @@
-<?php
+<?php declare(strict_types=1);
 /**
  * @package   Elabftw\Elabftw
  * @author    Nicolas CARPi <nico-git@deltablot.email>
@@ -6,7 +6,6 @@
  * @license   https://www.gnu.org/licenses/agpl-3.0.html AGPL-3.0
  * @see       https://www.elabftw.net Official website
  */
-declare(strict_types=1);
 
 namespace Elabftw\Interfaces;
 
@@ -25,4 +24,6 @@ interface ProcessorInterface
     public function getParams();
 
     public function getAction(): string;
+
+    public function getTarget(): string;
 }
